Add files via upload
Here you will find all the tasks that we aim to complete. If you choose to take one please announce that you have on the discord to alert others that the task is taken. Please aim for the lower-tiered tasks as they build into the higher ones. If you feel that more should be added/altered feel free to notify us on discord